create-modulemd: Resolve deps with common-build-dependencies as a "runtime" dependency
intltool and itstool are in flatpak-runtime, but not in the 'runtime' profile -
instead they are there for things that buildrequires flatpak-runtime.

intltool and itstool depend on autoconf and automake, but we don't want to build
them as part of flatpak-runtime; instead we want to pull those from
common-build-dependencies.

For this reason, we need to treat common-build-dependencies as a 'requires' of
the runtime.