Add module metadata for modules downloaded from Koji
As of Fedora 29, dnf enforces module dependencies, so to install an
application module that depends on flatpak-runtime, the DNF
repository with flatpak-runtime content needs the modulemd included.
However, MBS doesn't add the modulemd when downloading module
content from koji. So we create a second directory, symlink the
MBS-downloaded content into it, and run createrepo/modifyrepo again.