Pass --disable-sandbox to 'flatpak build-export'
Flatpak 1.2.0 adds validation of icons during the export process. The validation
of icons is run within a bubblewrap sandbox, which won't work within an
unprivileged container. So check for basic operation of bubblewrap sandboxes,
and if that fails, add the --disable-sandbox option to 'flatpak build-export'.
This is new in 1.2.1 so retry without --disable-sandbox if that fails.