FlatpakBuilder: Add end-of-life and end-of-life-rebase
end-of-life allows marking apps and runtimes as EOL. This is mostly
useful for runtimes where we can then mark older versions that are no
longer supported as EOL, so that they can be removed by client software.

end-of-life-rebase makes it possible to rename app IDs, so that users
get automatically migrated to the new ID. This commit wires it up for
runtimes as well, but I expect it won't get used there.