Add module_hotfixes=true to the repo files we create for buildrequires
Recent versions of dnf filter out packages with a modularitylabel RPM
tag, unless A) there is module metadata for an enabled module for them
or B) module_hotfixes is set on the repo. Set module_hotfixes on repos
for downloaded buildrequires to match what MBS is doing for local builds.