Disable bootstrap for mock builds
Recent mock versions default to use_bootstrap=True. The bootstrap process,
however, breaks with our filtered repositories, which don't contain dnf.
So disable it and assume that the host system is close enough to the RPMs
we are working with. We use the old name use_bootstrap_container for maximum
compatibility.