flatpak_builder: Fix component list for duplication between runtime and modules
If the Flatpak ended up with a package installed that had the same name as one
of the package from the runtime, then it would be omitted from the result
of get_components(), since we were filtering out packages by checking if their
name was in the runtime profile.

Instead of filtering out using the runtime profile, instead include only packages
with NEVR's in one of the app modules.