Allow overriding the architecture that FlatpakBuilder uses
We want to allow cross-building and generating a Flatpak OCI for
an architecture other than the current one - to enable this, allow
FlatpakBuilder(..., oci_arch=<other arch).

Because we need to maintain a mapping between OCI/RPM/Flatpak
architecture names, new architectures will require an addition here.

Tests are updated to handle metadata changes based on architecture.

We also change binary package names in the test data based on
architecture - this is not necessary, but the mismatch is quite
confusing if we don't do this.