flatpak_builder: error out cleanly on files in /usr/etc
According to comments in the code, we were meant to be dropping files
in /usr/etc (which conflicts with /etc with the Flatpak runtime files/
layer), but that wasn't actually implemented correctly.

Since files in /usr/etc typically reflect a packaging bug, error out
with a clean error message instead.