installer: Handle missing org.opencontainers.image.ref.name annotation
Flatpak expects an OCI toplevel index that it is installing from to have
the org.opencontainers.image.ref.name annotation on each image.
We've already parsed the ref name out, so write the annotation.
(This is necessary when the OCI toplevel is produced by buildah
or skopeo rather than Flatpak.)