cli: Add container-install and container-export commands
These commands are meant to be used when building from a
Containerfile. The content can be installed either using
container-install, or directly in the ContainerFile, then
container-export is used to make an OCI container from that.

Using a special feature of buildah (or podman build), we can
then make this OCI container the result of the Containerfile -
the final stages look something like:

  FROM flatpak-builder
  RUN --mount=type=bind,src=/,dst=/contents,from=install \
      --mount=type=bind,rw,z,src=export,dst=/export \
    flatpak-module container-export \
        --containerspec=/tmp/${containerspec} \
        --resultdir=/export

  FROM oci-archive:export/out.ociarchive