Chroot: Move handling of chroot setup to a separate class
The new Chroot class incorporates the logic for checking what type
of "unshare" we need to create a skeleton structure for the
chroot we install into. The logic is extended to handle the case
where we already have CAP_SYS_ADMIN and can create a new mount
namespace, but can't create a new user namespace.